requirements.txt
pandas
requests
logging
python-dotenv
